requirements.txt
psycopg2-binary>=2.9.9
notion-client>=3.0.0
python-dotenv>=1.0.0
